batch

+++ controllers/users/admin/delete.js
@@ -1,12 +1,11 @@
 const deleteDocument = require("../../../service/mainCRUD/deleteDoc");
 const deleteUser = require("../../../service/users/firebaseAuth/deleteUser");
-const deleteId1FieldAndReduceTotal = require("../../../service/utils/deleteId1FieldAndReduceTotal");
+// const deleteId1FieldAndReduceTotal = require("../..service/utils/deleteId1FieldAndReduceTotal");
 const getDocumentDataById = require("../../../service/utils/getDocumentDataById");
 const popArrayElement = require("../../../service/utils/popArrayElementFromObject");
 const updateOrCreateFieldsInDocument = require("../../../service/utils/updateOrCreateFieldsInDocument");
 const admin = require("../../../config/firebase-admin"); // Import Firebase Admin
-const disableUserAccount = require("../../../service/users/firebaseAuth/disableUser");
-const revokeRefreshTokens = require("../../../service/users/firebaseAuth/revokeRefreshTokens");
+const deleteId1FieldAndReduceTotal = require("../../../service/utils/deleteId1FieldAndReduceTotal");
 
 /**
  * Delete an admin document from the "admin" Firestore collection.
@@ -28,6 +27,9 @@ const deleteAdmin = async (req, res) => {
     const db = admin.firestore();
     const batch = db.batch();
 
+    await updateOrCreateFieldsInDocument(db, batch, "admin", id, {
+      disable: true,
+    });
     // Step 3: Delete the admin document from the "admin" collection
     await deleteDocument(db, batch, "admin", id);
 
@@ -55,31 +57,26 @@ const deleteAdmin = async (req, res) => {
       db,
       batch
     );
-    await popArrayElement(
-      "worker",
-      { id: id },
-      adminData.branchId,
-      "branches",
-      db,
-      batch
-    );
-    const branchData = await getDocumentDataById(
-      "branches",
-      adminData.branchId
-    );
-    if (branchData && branchData.active) {
-      await deleteId1FieldAndReduceTotal(branchData.active, id, db, batch);
-    }
 
+    if (adminData.branchId) {
+      await popArrayElement(
+        "worker",
+        { id: id },
+        adminData.branchId,
+        "branches",
+        db,
+        batch
+      );
+      const branchData = await getDocumentDataById(
+        "branches",
+        adminData.branchId
+      );
+      if (branchData && branchData.active) {
+        await deleteId1FieldAndReduceTotal(branchData.active, id, db, batch);
+      }
+    }
     // Step 4: Update or create the "disable" field in the specified Firestore collection document
-    await updateOrCreateFieldsInDocument(db, batch, collectionName, id, {
-      disable: true,
-    });
-    // Step 2: Disable or enable the user account in Firebase Authentication
-    await disableUserAccount(id, true);
-    // Step 3: If disabled, revoke the user's refresh tokens
-    await revokeRefreshTokens(id);
-    await deleteUser(id);
+
     await batch.commit();
     // Step 7: Respond with a success message
     res.status(200).json({ message: "Admin document deleted successfully." });
